Module async_raft::raft

Public Raft interface and data types.

Structs

AppendEntriesRequest

An RPC sent by a cluster leader to replicate log entries (5.3), and as a heartbeat (5.2).

AppendEntriesResponse

The response to an AppendEntriesRequest.

ClientWriteRequest

An application specific client request to update the state of the system (5.1).

ClientWriteResponse

The response to a ClientRequest.

ConflictOpt

A struct used to implement the conflicting term optimization outlined in 5.3 for log replication.

Entry

A Raft log entry.

EntryConfigChange

A log entry holding a config change.

EntryNormal

A normal log entry.

EntrySnapshotPointer

A log entry pointing to a snapshot.

InstallSnapshotRequest

An RPC sent by the Raft leader to send chunks of a snapshot to a follower (7).

InstallSnapshotResponse

The response to an InstallSnapshotRequest.

MembershipConfig

A model of the membership configuration of the cluster.

Raft

The Raft API.

VoteRequest

An RPC sent by candidates to gather votes (5.2).

VoteResponse

The response to a VoteRequest.

Enums

EntryPayload

Log entry payload variants.